Cleanup

+++ src/CodeGenerator/FileWriter.php
@@ -56,6 +56,8 @@ class FileWriter
         // Remove AsyncAws\
         $class = substr($class, 9);
         $file = \sprintf('%s/%s.php', $this->srcDirectory, str_replace('\\', '/', $class));
-        unlink($file);
+        if (is_file($file)) {
+            unlink($file);
+        }
     }
 }
